Properly load data on big screens when changing node
Previously, the app would only fetch the new data on component mount.
Since the `NodeDetails` component is not re-mounted when the screen is
big enough (it will always be in view), it would not get updated props.

This commit converts the component to a functional component and then
adds a `useEffect` that will react to any changes in the `match` prop
provided by react-router.